Factor out new qwikstart.utils.dict_utils
