+++ qwikstart/utils/__init__.py
@@ -1,11 +1,5 @@
-from .core import (
-    ensure_path,
-    first,
-    full_class_name,
-    indent,
-    merge_nested_dicts,
-    remap_dict,
-)
+from .core import ensure_path, first, full_class_name, indent
+from .dict_utils import merge_nested_dicts, remap_dict
 
 __all__ = [
     "ensure_path",
